fixing php errors

+++ modules/InboundEmail/InboundEmail.php
@@ -680,7 +680,12 @@ class InboundEmail extends SugarBean
         }
         $this->connectMailserver();
 
-        $uids = imap_search($this->conn, "ALL", SE_UID);
+        if (is_resource($this->conn)) {
+            $uids = imap_search($this->conn, "ALL", SE_UID);
+        } else {
+            LoggerManager::getLogger()->warn('connection is not a valid resource to empty trush');
+            $uids = array();
+        }
 
         foreach ($uids as $uid) {
             if (!imap_delete($this->conn, $uid, FT_UID)) {
@@ -1083,7 +1088,12 @@ class InboundEmail extends SugarBean
                             if (isset($overview->uid) && !empty($overview->uid)) {
                                 $this->imap_uid = $overview->uid;
                             }
-                            $values .= "'{$this->imap_uid}'";
+                            if (!isset($this->imap_uid)) {
+                                LoggerManager::getLogger()->warn('Inbound email has not imap uid for setting cache value.');
+                                $values .= "''";
+                            } else {
+                                $values .= "'{$this->imap_uid}'";
+                            }
                             break;
 
                         case "ie_id":
@@ -1636,7 +1646,12 @@ class InboundEmail extends SugarBean
         }
         $this->setCacheTimestamp($mailbox);
         $GLOBALS['log']->info("[EMAIL] Performing IMAP search using criteria [{$criteria}] on mailbox [{$mailbox}] for user [{$current_user->user_name}]");
-        $searchResults = imap_search($this->conn, $criteria, SE_UID);
+        if (!is_resource($this->conn)) {
+            LoggerManager::getLogger()->warn('checkEmailOneMailbox: connection is not a valid resource');
+            $searchResults = null;
+        } else {
+            $searchResults = imap_search($this->conn, $criteria, SE_UID);
+        }
         $GLOBALS['log']->info("[EMAIL] Done IMAP search on mailbox [{$mailbox}] for user [{$current_user->user_name}]. Result count = " . count($searchResults));
 
         if (!empty($searchResults)) {
@@ -1676,7 +1691,12 @@ class InboundEmail extends SugarBean
         }
 
         if ($this->mailbox != $trashFolder) {
-            $searchResults = imap_search($this->conn, $criteria, SE_UID);
+            if (!is_resource($this->conn)) {
+                LoggerManager::getLogger()->warn('connection is not a valid resource for checkEmailOneMailbox()');
+                $searchResults = null;
+            } else {
+                $searchResults = imap_search($this->conn, $criteria, SE_UID);
+            }
             if (!empty($searchResults)) {
                 $uids = implode($app_strings['LBL_EMAIL_DELIMITER'], $searchResults);
                 $GLOBALS['log']->info("INBOUNDEMAIL: removing UIDs found deleted [ {$uids} ]");
@@ -1762,7 +1782,9 @@ class InboundEmail extends SugarBean
         if ($ret['status'] == 'done') {
             //Remove the cached search if we are done with this mailbox
             $cacheFilePath = clean_path("{$this->EmailCachePath}/{$this->id}/folders/SearchData.php");
-            unlink($cacheFilePath);
+            if (file_exists($cacheFilePath)) {
+                unlink($cacheFilePath);
+            }
             /**
              * To handle the use case where an external client is also connected, deleting emails, we need to clear our
              * local cache of all emails with the "DELETED" flag
@@ -1777,7 +1799,12 @@ class InboundEmail extends SugarBean
             }
 
             if ($this->mailbox != $trashFolder) {
-                $searchResults = imap_search($this->conn, $criteria, SE_UID);
+                if (!is_resource($this->conn)) {
+                    LoggerManager::getLogger()->warn('mailbox != trash folder but connection is not a valid resource for checkEmailOneMailbox()');
+                    $searchResults = null;
+                } else {
+                    $searchResults = imap_search($this->conn, $criteria, SE_UID);
+                }
                 if (!empty($searchResults)) {
                     $uids = implode($app_strings['LBL_EMAIL_DELIMITER'], $searchResults);
                     $GLOBALS['log']->info("INBOUNDEMAIL: removing UIDs found deleted [ {$uids} ]");
@@ -1903,6 +1930,9 @@ class InboundEmail extends SugarBean
         $ret['mailboxes'] = $mailboxes_meta;
 
         foreach ($mailboxes_meta as $count) {
+            if (!isset($ret['processCount'])) {
+                $ret['processCount'] = 0;
+            }
             $ret['processCount'] += $count;
         }
 
@@ -1934,7 +1964,13 @@ class InboundEmail extends SugarBean
         }
 
         $GLOBALS['log']->info("INBOUNDEMAIL: using [ {$criteria} ]");
-        $searchResults = imap_search($this->conn, $criteria, SE_UID);
+        if (!is_resource($this->conn)) {
+            LoggerManager::getLogger()->warn('connection is not a valid resource for getMailboxProcessCount()');
+            $searchResults = null;
+        } else {
+            $searchResults = imap_search($this->conn, $criteria, SE_UID);
+        }
+        
 
         if (!empty($searchResults)) {
             $concatResults = implode(",", $searchResults);
@@ -1961,7 +1997,12 @@ class InboundEmail extends SugarBean
         } else {
             $this->connectMailserver();
             $mailboxes = $this->getMailboxes(true);
-            sort($mailboxes);
+            if (!is_array($mailboxes)) {
+                LoggerManager::getLogger()->warn('mailboxes is not an array for check email');
+                $mailboxes = (array)$mailboxes;
+            } else {
+                sort($mailboxes);
+            }
 
             $GLOBALS['log']->info("INBOUNDEMAIL: checking account [ {$this->name} ]");
 
@@ -6910,6 +6951,11 @@ class InboundEmail extends SugarBean
         $r = $this->db->query($query);
         $a = $this->db->fetchByAssoc($r);
 
+        if (!isset($a['message_id'])) {
+            LoggerManager::getLogger()->warn('unable to get msgno for message id');
+            return null;
+        }
+        
         return $a['message_id'];
     }
